refactor: extract keepalive port logic into RuntimeService.connectPort
popup/sidepanel で重複していたポート接続・ping・再接続のロジックを
runtimeService.connectPort() に集約する。
呼び出し側は1行に簡略化され、keepalive の詳細はサービス層に隠蔽される。
