+++ src/entrypoints/popup/main.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { contentService } from "@/services/content";
+import { runtimeService } from "@/services/runtime";
 import App from "./App";
 
 const rootElement =
@@ -12,23 +13,8 @@ chrome.runtime.onMessage.addListener((message) => {
   }
 });
 
-// MV3 Service Worker はアイドル時に停止されるため、ポート接続で生存を維持する。
-// ポップアップが開いている間は SW を生かし続け、メッセージ送信の失敗を防ぐ。
-const KEEPALIVE_INTERVAL_MS = 25000;
-
-function connectKeepalivePort(): void {
-  const port = chrome.runtime.connect({ name: "keepalive" });
-  const interval = setInterval(
-    () => port.postMessage({ type: "PING" }),
-    KEEPALIVE_INTERVAL_MS,
-  );
-  port.onDisconnect.addListener(() => {
-    clearInterval(interval);
-    connectKeepalivePort();
-  });
-}
-
-connectKeepalivePort();
+// MV3 Service Worker のアイドル停止を防ぐ永続ポート接続を確立する
+runtimeService.connectPort("keepalive");
 
 ReactDOM.createRoot(rootElement).render(
   <React.StrictMode>
